+++ lib/ui/pages/holder/game_center/user_prediction_page/widget/user_prediction_card_list.dart
@@ -1,4 +1,4 @@
-import 'package:ballkkaye_frontend/_core/utils/m_time.dart';
+import 'package:ballkkaye_frontend/_core/utils/m_util.dart';
 import 'package:ballkkaye_frontend/ui/pages/holder/game_center/user_prediction_page/user_prediction_vm.dart';
 import 'package:ballkkaye_frontend/ui/pages/holder/game_center/user_prediction_page/widget/user_prediction_card.dart';
 import 'package:flutter/material.dart';
